Example works but code not complete

+++ inst/include/result.h
@@ -59,7 +59,7 @@ struct rejection_result
 	std::vector<unsigned int> rejects;
 	std::vector<unsigned int> tunes;
 	std::vector<unsigned int> regions;
-	std::vector<unsigned int> log_bounds;
+	std::vector<double> log_bounds;
 
 	operator SEXP() const;
 };
